[TSK-41] Feat: 판매 게시글 CRUD 구현

+++ src/main/kotlin/com/wafflestudio/team2server/area/controller/AreaController.kt
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.RestController
 @RestController
 class AreaController(private val areaService: AreaService) {
 	@GetMapping("/area")
-	fun getAreaInfo(code: Long) {
+	fun getAreaInfo(code: Int) {
 
 	}
 }
