+++ src/main/kotlin/com/wafflestudio/team2server/area/repository/AreaAdjRepository.kt
@@ -4,5 +4,5 @@ import com.wafflestudio.team2server.area.model.AreaAdjEntity
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface AreaAdjRepository : JpaRepository<AreaAdjEntity, Long> {
-	fun getAreaAdjEntityByAreaIdAndDistanceIsLessThanEqual(id: Long, distance: Int): List<AreaAdjEntity>
+	fun getAreaAdjEntityByAreaIdAndDistanceIsLessThanEqual(id: Int, distance: Int): List<AreaAdjEntity>
 }
