+++ src/main/kotlin/com/wafflestudio/team2server/area/repository/AreaRepository.kt
@@ -3,5 +3,5 @@ package com.wafflestudio.team2server.area.repository
 import com.wafflestudio.team2server.area.model.AreaEntity
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface AreaRepository: JpaRepository<AreaEntity, Long> {
+interface AreaRepository : JpaRepository<AreaEntity, Int> {
 }
